src/oneid/oneid-lambda-cognito-proxy/requirements.txt
boto3==1.38.1
python-jose==3.4.0
requests==2.32.3
